Move test doubles into their own file
