+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/News/Compositional/Widgets/Events/Test Doubles/FakeEventsWidgetViewModel.swift
@@ -1,0 +1,24 @@
+import EurofurenceApplication
+
+class FakeEventsWidgetViewModel: EventsWidgetViewModel {
+    
+    typealias Event = FakeEventViewModel
+    
+    var title: String = ""
+    
+    var numberOfEvents: Int {
+        events.count
+    }
+    
+    var events: [FakeEventViewModel] = []
+    
+    func event(at index: Int) -> Event {
+        events[index]
+    }
+    
+    private(set) var selectedEventIndex: Int?
+    func eventSelected(at index: Int) {
+        selectedEventIndex = index
+    }
+    
+}
