+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/News/Compositional/Widgets/Events/Test Doubles/FakeEventsWidgetViewModelFactory.swift
@@ -1,0 +1,12 @@
+import EurofurenceApplication
+import RouterCore
+
+struct FakeEventsWidgetViewModelFactory: NewsWidgetViewModelFactory {
+    
+    let viewModel: FakeEventsWidgetViewModel
+    
+    func makeViewModel(router: Router) -> some EventsWidgetViewModel {
+        viewModel
+    }
+    
+}
