+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/News/Compositional/Widgets/Events/Test Doubles/FakeNewsWidgetManager.swift
@@ -1,0 +1,21 @@
+import EurofurenceApplication
+import RouterCore
+import UIKit
+
+class FakeNewsWidgetManager: NewsWidgetEnvironment {
+    
+    private let tableView: UITableView
+    private(set) var installedDataSources = [TableViewMediator]()
+    
+    init(tableView: UITableView) {
+        self.tableView = tableView
+    }
+    
+    var router: Router = Routes()
+    
+    func install(dataSource: TableViewMediator) {
+        installedDataSources.append(dataSource)
+        dataSource.registerReusableViews(into: tableView)
+    }
+    
+}
