+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/News/Compositional/Widgets/Events/View/EventsNewsWidgetTableViewDataSourceTests.swift
@@ -305,102 +305,6 @@ class EventsNewsWidgetTableViewDataSourceTests: XCTestCase {
         XCTAssertEqual(expected, actual)
     }
     
-    private class FakeNewsWidgetManager: NewsWidgetEnvironment {
-        
-        private let tableView: UITableView
-        private(set) var installedDataSources = [TableViewMediator]()
-        
-        init(tableView: UITableView) {
-            self.tableView = tableView
-        }
-        
-        var router: Router = Routes()
-        
-        func install(dataSource: TableViewMediator) {
-            installedDataSources.append(dataSource)
-            dataSource.registerReusableViews(into: tableView)
-        }
-        
-    }
-    
-    private struct FakeEventsWidgetViewModelFactory: NewsWidgetViewModelFactory {
-        
-        let viewModel: FakeEventsWidgetViewModel
-        
-        func makeViewModel(router: Router) -> some EventsWidgetViewModel {
-            viewModel
-        }
-        
-    }
-    
-    private class FakeEventsWidgetViewModel: EventsWidgetViewModel {
-        
-        typealias Event = FakeEventViewModel
-        
-        var title: String = ""
-        
-        var numberOfEvents: Int {
-            events.count
-        }
-        
-        var events: [FakeEventViewModel] = []
-        
-        func event(at index: Int) -> Event {
-            events[index]
-        }
-        
-        private(set) var selectedEventIndex: Int?
-        func eventSelected(at index: Int) {
-            selectedEventIndex = index
-        }
-        
-    }
-    
-    private class FakeEventViewModel: EventViewModel {
-        
-        @Observed var name: String
-        @Observed var location: String
-        @Observed var startTime: String
-        @Observed var endTime: String
-        @Observed var isFavourite: Bool
-        @Observed var isSponsorOnly: Bool
-        @Observed var isSuperSponsorOnly: Bool
-        @Observed var isArtShow: Bool
-        @Observed var isKageEvent: Bool
-        @Observed var isDealersDen: Bool
-        @Observed var isMainStage: Bool
-        @Observed var isPhotoshoot: Bool
-        
-        init(
-            name: String = "Name",
-            location: String = "Location",
-            startTime: String = "Start Time",
-            endTime: String = "End Time",
-            isFavourite: Bool = false,
-            isSponsorOnly: Bool = false,
-            isSuperSponsorOnly: Bool = false,
-            isArtShow: Bool = false,
-            isKageEvent: Bool = false,
-            isDealersDen: Bool = false,
-            isMainStage: Bool = false,
-            isPhotoshoot: Bool = false
-        ) {
-            self.name = name
-            self.location = location
-            self.startTime = startTime
-            self.endTime = endTime
-            self.isFavourite = isFavourite
-            self.isSponsorOnly = isSponsorOnly
-            self.isSuperSponsorOnly = isSuperSponsorOnly
-            self.isArtShow = isArtShow
-            self.isKageEvent = isKageEvent
-            self.isDealersDen = isDealersDen
-            self.isMainStage = isMainStage
-            self.isPhotoshoot = isPhotoshoot
-        }
-        
-    }
-    
 }
 
 private extension UIView {
